add paradigm30 chess
Paradigm30 chess variant, with 'b' for the bishop instead of 'd'. Since fairyland is a variant I personally made, rather than a public variant, it probably wouldn't fit on this website. Please respond if you would like me to add paradigm1320, which is on hold as it requires many changes to other files.